fix: adding an option to load analytics eager
