+++ scripts/delayed.js
@@ -1,6 +1,9 @@
 // eslint-disable-next-line import/no-cycle
 import { loadScript, sampleRUM } from './aem.js';
-import { getEnvType } from './blocks-utils.js';
+import {
+  // eslint-disable-next-line import/named
+  loadGTM, loadAdobeLaunch, loadAnalyticsEager,
+} from './blocks-utils.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
@@ -26,43 +29,17 @@ loadScript('https://www.google.com/recaptcha/api.js?onload=onCaptchaloadCallback
 /**
  * Google Tag Manager
 * */
-async function loadGTM() {
-  const scriptTag = document.createElement('script');
-  scriptTag.innerHTML = `
-        (function (w, d, s, l, i) {
-        w[l] = w[l] || [];
-        w[l].push({
-            'gtm.start':
-                new Date().getTime(), event: 'gtm.js'
-        });
-        var f = d.getElementsByTagName(s)[0],
-            j = d.createElement(s), dl = l != 'dataLayer' ? '&l=' + l : '';
-        j.async = true;
-        j.src =
-            'https://www.googletagmanager.com/gtm.js?id=' + i + dl;
-        f.parentNode.insertBefore(j, f);
-        }(window, document, 'script', 'dataLayer', 'GTM-WF9LTLZ'));
-    `;
-  document.head.prepend(scriptTag);
-}
 
 loadScript('/scripts/cookie-script.js');
 
 if (!isSidekickLibrary) {
-  // TODO: Remove delayed loading of GTM once it stops impacting page performance
-  setTimeout(() => {
-    loadGTM();
-  }, 2000);
+  if (!loadAnalyticsEager()) {
+    // TODO: Remove delayed loading of GTM once it stops impacting page performance
+    setTimeout(() => {
+      loadGTM();
+    }, 2000);
 
-  (function loadAdobeLaunch() {
-    const adobeLaunchSrc = {
-      dev: 'https://assets.adobedtm.com/64c36731dbac/390f7bab5b74/launch-285ee83071cc-development.min.js',
-      preview: 'https://assets.adobedtm.com/64c36731dbac/390f7bab5b74/launch-285ee83071cc-development.min.js',
-      live: 'https://assets.adobedtm.com/64c36731dbac/390f7bab5b74/launch-285ee83071cc-development.min.js',
-      prod: 'https://assets.adobedtm.com/64c36731dbac/390f7bab5b74/launch-285ee83071cc-development.min.js',
-    };
-    loadScript(adobeLaunchSrc[getEnvType()], { async: true });
-  }());
-
+    loadAdobeLaunch();
+  }
   loadScript('https://icici-securities.allincall.in/files/deploy/embed_chatbot_11.js?version=1.1');
 }
